refactor: migrate type imports from @/lib/api to @/types (#321)
Complete migration of all type imports to use @/types as canonical source,
removing the transitional re-export layer from api.ts introduced in PR #320.

Changes:
- Migrate 39 files to import types from @/types instead of @/lib/api
- Split mixed imports: types from @/types, functions from @/lib/api
- Remove re-export layer (lines 5-112) from frontend/src/lib/api.ts
- Update api.ts to import types from @/types for internal use

Files modified:
- 15 component files
- 10 test files
- 5 hook files
- 2 page files
- 1 utility file (tfl-colors.ts)
- 1 context file (BackendAuthContext.tsx)
- 5 SegmentBuilder module files

Verification:
- TypeScript compilation passes
- All tests pass (517 passed, 5 skipped)
- Frontend build succeeds
- Zero type imports from @/lib/api remain

Closes #321
