+++ frontend/src/components/contacts/VerificationDialog.tsx
@@ -12,7 +12,7 @@ import { Button } from '../ui/button'
 import { Input } from '../ui/input'
 import { Label } from '../ui/label'
 import { Alert, AlertDescription } from '../ui/alert'
-import type { Contact } from '../../lib/api'
+import type { Contact } from '@/types'
 
 export interface VerificationDialogProps {
   open: boolean
