+++ frontend/src/components/routes/LineButton.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { LineButton } from './LineButton'
-import type { LineResponse } from '@/lib/api'
+import type { LineResponse } from '@/types'
 import { CORPORATE_BLUE, WHITE } from '@/lib/tfl-colors'
 
 describe('LineButton', () => {
