+++ frontend/src/lib/api.ts
@@ -2,114 +2,45 @@
  * API client for backend communication
  */
 
-// Import generated API types for local use
 import type {
   UserResponse,
+  ContactsResponse,
   EmailResponse,
   PhoneResponse,
-  ContactsResponse,
   AddEmailRequest,
   AddPhoneRequest,
-  VerifyCodeRequest,
   SendVerificationResponse,
+  VerifyCodeRequest,
   VerifyCodeResponse,
-  SegmentResponse,
-  ScheduleResponse,
-  RouteResponse,
   RouteListItemResponse,
+  RouteResponse,
   CreateRouteRequest,
   UpdateRouteRequest,
+  SegmentResponse,
   SegmentRequest,
   UpdateSegmentRequest,
   CreateScheduleRequest,
+  ScheduleResponse,
   UpdateScheduleRequest,
   NotificationPreferenceResponse,
   CreateNotificationPreferenceRequest,
-  NotificationMethod,
-  NotificationStatus,
-  Contact,
-  NetworkGraph,
-  RouteVariant,
   LineResponse,
   StationResponse,
-  NetworkConnection,
+  NetworkGraph,
   RouteValidationSegment,
   RouteValidationResponse,
   RebuildIndexesResponse,
-  SyncMetadataResponse,
   BuildGraphResponse,
+  SyncMetadataResponse,
   TriggerCheckResponse,
   WorkerStatusResponse,
-  NotificationLogItem,
+  NotificationStatus,
   RecentLogsResponse,
-  EmailAddressItem,
-  PhoneNumberItem,
-  UserListItem,
   PaginatedUsersResponse,
   UserDetailResponse,
   AnonymiseUserResponse,
-  UserCountMetrics,
-  RouteStatMetrics,
-  NotificationStatMetrics,
-  DailySignup,
-  GrowthMetrics,
   EngagementMetrics,
 } from '@/types'
-
-// Re-export types for backward compatibility
-// Files currently import from '@/lib/api', so we re-export types from '@/types'
-export type {
-  UserResponse,
-  EmailResponse,
-  PhoneResponse,
-  ContactsResponse,
-  AddEmailRequest,
-  AddPhoneRequest,
-  VerifyCodeRequest,
-  SendVerificationResponse,
-  VerifyCodeResponse,
-  SegmentResponse,
-  ScheduleResponse,
-  RouteResponse,
-  RouteListItemResponse,
-  CreateRouteRequest,
-  UpdateRouteRequest,
-  SegmentRequest,
-  UpdateSegmentRequest,
-  CreateScheduleRequest,
-  UpdateScheduleRequest,
-  NotificationPreferenceResponse,
-  CreateNotificationPreferenceRequest,
-  NotificationMethod,
-  NotificationStatus,
-  Contact,
-  NetworkGraph,
-  RouteVariant,
-  LineResponse,
-  StationResponse,
-  NetworkConnection,
-  RouteValidationSegment,
-  RouteValidationResponse,
-  RebuildIndexesResponse,
-  SyncMetadataResponse,
-  BuildGraphResponse,
-  TriggerCheckResponse,
-  WorkerStatusResponse,
-  NotificationLogItem,
-  RecentLogsResponse,
-  EmailAddressItem,
-  PhoneNumberItem,
-  UserListItem,
-  PaginatedUsersResponse,
-  UserDetailResponse,
-  AnonymiseUserResponse,
-  UserCountMetrics,
-  RouteStatMetrics,
-  NotificationStatMetrics,
-  DailySignup,
-  GrowthMetrics,
-  EngagementMetrics,
-}
 
 /**
  * Module-level API base URL
