+++ frontend/src/pages/admin/AdminUsers.test.tsx
@@ -2,7 +2,8 @@ import { render, screen, waitFor, within } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import AdminUsers from './AdminUsers'
-import { ApiError, type PaginatedUsersResponse, type UserDetailResponse } from '@/lib/api'
+import type { PaginatedUsersResponse, UserDetailResponse } from '@/types'
+import { ApiError } from '@/lib/api'
 
 // Mock the useAdminUsers hook
 vi.mock('@/hooks/useAdminUsers', () => ({
